requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
asttokens=2.2.1=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
brotlipy=0.7.0=py39hb9d737c_1005
bzip2=1.0.8=h7f98852_4
ca-certificates=2023.7.22=hbcca054_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py39he91dace_3
charset-normalizer=3.1.0=pyhd8ed1ab_0
click=8.0.4=py39h06a4308_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.3=pyhd8ed1ab_0
contourpy=1.1.1=pypi_0
cryptography=40.0.1=py39h079d5ae_0
cudatoolkit=11.1.1=ha002fc5_11
cycler=0.12.1=pypi_0
debugpy=1.6.7=py39h227be39_0
decorator=5.1.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
ffmpeg=4.3=hf484d3e_0
filelock=3.11.0=pyhd8ed1ab_0
fonttools=4.43.1=pypi_0
freetype=2.12.1=hca18f0e_1
gin-config=0.5.0=pyhd8ed1ab_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
huggingface_hub=0.13.4=pyhd8ed1ab_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.3.0=pyha770c72_0
importlib-resources=6.1.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.23.1=pyh210e3f2_0
ipython=8.14.0=pyh41d4057_0
jedi=0.18.2=pyhd8ed1ab_0
joblib=1.1.1=py39h06a4308_0
jpeg=9b=h024ee3a_2
jupyter_client=7.4.9=pyhd8ed1ab_0
jupyter_core=4.12.0=py39hf3d152e_0
kiwisolver=1.4.5=pypi_0
lame=3.100=h166bdaf_1003
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=h41732ed_0
libffi=3.4.2=h7f98852_5
libgcc-ng=12.2.0=h65d4601_19
libgfortran-ng=13.2.0=h69a702a_0
libgfortran5=13.2.0=ha4646dd_0
libgomp=12.2.0=h65d4601_19
libiconv=1.17=h166bdaf_0
libnsl=2.0.0=h7f98852_0
libpng=1.6.39=h753d276_0
libsodium=1.0.18=h36c2ea0_1
libsqlite=3.40.0=h753d276_0
libstdcxx-ng=12.2.0=h46fd767_19
libtiff=4.2.0=h85742a9_0
libuuid=2.38.1=h0b41bf4_0
libwebp-base=1.3.0=h0b41bf4_0
libzlib=1.2.13=h166bdaf_4
lz4-c=1.9.3=h9c3ff4c_1
matplotlib=3.8.1=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7e14d7c_0
mkl_fft=1.3.1=py39h0c7bc48_1
mkl_random=1.2.2=py39hde0f152_0
ncurses=6.3=h27087fc_1
nest-asyncio=1.5.6=pyhd8ed1ab_0
nettle=3.6=he412f7d_0
ninja=1.11.1=h924138e_0
nltk=3.7=pyhd3eb1b0_0
numpy=1.23.5=py39h14f4228_0
numpy-base=1.23.5=py39h31eccc5_0
olefile=0.46=pyh9f0ad1d_1
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=hb52868f_1
openssl=3.1.4=hd590300_0
packaging=23.0=pyhd8ed1ab_0
pandas=2.0.0=py39h2ad29b5_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=8.3.1=py39h2c7a002_0
pip=23.0.1=pyhd8ed1ab_0
prompt-toolkit=3.0.38=pyha770c72_0
prompt_toolkit=3.0.38=hd8ed1ab_0
psutil=5.9.5=py39h72bdee0_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.15.1=pyhd8ed1ab_0
pyopenssl=23.1.1=pyhd8ed1ab_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.9.16=h2782a2a_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.9=3_cp39
pytorch=1.8.0=py3.9_cuda11.1_cudnn8.0.5_0
pytz=2023.3=pyhd8ed1ab_0
pyyaml=6.0=py39hb9d737c_5
pyzmq=25.1.0=py39hb257651_0
readline=8.2=h8228510_1
regex=2023.3.23=py39h72bdee0_0
requests=2.28.2=pyhd8ed1ab_1
scikit-learn=1.3.2=py39ha22ef79_1
scipy=1.9.3=py39h14f4228_1
seqeval=1.2.2=pyhd3deb0d_0
setuptools=67.6.1=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
stack_data=0.6.2=pyhd8ed1ab_0
threadpoolctl=3.2.0=pyha21a80b_0
tk=8.6.12=h27826a3_0
tokenizers=0.13.3=py39h585fa2d_0
torchaudio=0.8.0=py39
torchvision=0.9.0=py39_cu111
tornado=6.3.2=py39hd1e30aa_0
tqdm=4.65.0=pyhd8ed1ab_1
traitlets=5.9.0=pyhd8ed1ab_0
transformers=4.24.0=py39h06a4308_0
typing-extensions=4.5.0=hd8ed1ab_0
typing_extensions=4.5.0=pyha770c72_0
tzdata=2023c=h71feb2d_0
urllib3=1.26.15=pyhd8ed1ab_0
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.40.0=pyhd8ed1ab_0
xz=5.2.6=h166bdaf_0
yaml=0.2.5=h7f98852_2
zeromq=4.3.4=h9c3ff4c_1
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h166bdaf_4
zstd=1.4.9=ha95c52a_0
